+++ src/index.ts
@@ -1,0 +1,5 @@
+import Fritz from './Fritz';
+
+export * from './utils';
+export { Fritz };
+export default Fritz;